Slack_automation/slack-html-reports/requirements.txt
slack-sdk
jinja2
python-dotenv
schedule==1.2.2